# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Todo API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Login for registered users",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Login for registered users",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Signup for new users",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Signup for new users",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserSignup"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/todos/": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Get all the logged in user's todos with the category associated with each todo and status that is not the deleted status",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "description": "Add a new todo with their category for the logged in user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Add a new todo with their category for the logged in user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewTodoWithCategory"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/todos/all": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Get all the users todos including todos with a deleted status",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/todos/deleted": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Get all the todos with the deleted status",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/todos/statuses": {
      "get": {
        "tags": [
          "Status"
        ],
        "description": "Get all the statuses from the database",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/todos/{id}": {
      "put": {
        "tags": [
          "Todo"
        ],
        "description": "Change/update a specific todo for logged in user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "statusName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "description": "Delete a specific todo if for the logged in user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/statuses/": {
      "post": {
        "tags": [
          "Status"
        ],
        "description": "Add new status to do database",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "statusName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/statuses/{id}": {
      "delete": {
        "tags": [
          "Status"
        ],
        "description": "Delete a status by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Status"
        ],
        "description": "Update the name of a status by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
    "UserSignup": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "johndoe@yahoo.com"
        },
        "password": {
          "type": "string",
          "example": "0000"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ]
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "johndoe@yahoo.com"
        },
        "password": {
          "type": "string",
          "example": "0000"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "NewTodoWithCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "todo 1"
        },
        "description": {
          "type": "string",
          "example": "description todo"
        },
        "categoryName": {
          "type": "string",
          "example": "category 1"
        }
      },
      "required": [
        "name",
        "description",
        "categoryName"
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}